Fix the enable for the PC_ACC: it must be on (0) when either there is not a bubble (bubble=1) or  there is a overwrite.
Note that in this case overwrite is done also if the current instruction is a bubble (good),
but a bubble can never generate an overwrite (as a bubble sets no_check to 1).

Also, following the principle we said for timing, it is fixed as follows:

As PC_ACC is clocked, the signals changing on a clock edge (FALLBACK_ADDRESS and OVERWRITE) are latched to honour hold times.

Following the