Fallback address was computed in a bad way.
We didn't recognize this before because we had a nottaken BPU (PC did go on regularly) and PC arrived one stage early into the INP1 port of datapath (so it was
in fact PC+4). Fixed.

Now, the problem is that CU must compute PC+4 or PC+IMM+4.
In both cases, request is executed by means of a subtraction.
The input of select_rightB when S="00" (unused) has been changed to -4, so that result=PC-(-4)=PC+4 is case of predict taken.

In case of predict not taken, things are more difficult, as we have to compute PC+IMM+4;
we can exploit the fact that IMM is divisible by 4, so we compute -IMM-4 as (-(IMM/4)-1)*4, which in logic terms is equal to not(IMM(15 downto 2)) & "00",
as not(IMM(15 downto2))==-IMM/4-1.
In this case result=PC-(-IMM-4)=PC+IMM+4.

Well done, kids :)